requirements.txt
flask==3.0.0
werkzeug==3.0.1
numpy==1.26.2
opencv-python-headless==4.8.1.78
face_recognition==1.3.0
dlib==19.24.2
Pillow==10.1.0
gspread==5.12.0
google-auth==2.23.0
python-dotenv==1.0.0
setuptools<70.0.0